test-cases/form-test.html

<!DOCTYPE HTML>
<!-- 
Load a form into an iframe. The form auto submits, sending its form
data to node-server.js.  The server sends a message back that contains
JSON for the request, including headers, URL, method, and the 'body'
    
testharness.js is used to check the message sent back from the server
against required values.  The results of this comparison are displayed
using the div with id="log".
-->

<html>
  <head>
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
    <script src="form-test.js"></script>
  </head>

  <body>
    <h2>Various Form Tests</h2>
    <iframe id="form-frame" width="50%"></iframe>
    <div id="log">
      Log goes here
    </div>

    <script>			
    
// sample to show testing works
test(function() {assert_true(true)}, "assert_true with true")

// here's the real test
var test = async_test("Verify multipart/form-data Data");

test.step(function () {
    function receiveMessage(e) {
	var parsedParts = form_test.processResponse(JSON.parse(e.data));
	assert_object_equals(parsedParts, 
			     [{fieldName: "first-field", body: "Field One Text"},
			      {fieldName: "second-field", body: "Field Two Text"}],
			     "Successful multipart test");
	test.done(); 
    }
    window.addEventListener("message", test.step_func(receiveMessage));
    var formFrame = document.getElementById("form-frame");
    formFrame.src = "sub-form-test.html";
}); 
    </script>
  </body>
</html>
